+++ x68k/mfp.h
@@ -36,7 +36,7 @@ extern uint8_t LastKey;
 void MFP_Init(void);
 uint8_t FASTCALL MFP_Read(uint32_t adr);
 void FASTCALL MFP_Write(uint32_t adr, uint8_t data);
-void FASTCALL MFP_Timer(long clock);
+void FASTCALL MFP_Timer(int32_t clock);
 void FASTCALL MFP_TimerA(void);
 void MFP_Int(int irq);
